change invite tests to use members section instead of organization setting
